% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recordio.R
\name{recordio_close}
\alias{recordio_close}
\title{Close a recordio reader.}
\usage{
recordio_close(r)
}
\arguments{
\item{r}{reader object created by recordio_newreader.}
}
\value{
nil
}
\description{
Close a recordio reader.
}
