% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recordio.R
\name{recordio_newreader}
\alias{recordio_newreader}
\title{Create a recordio reader.}
\usage{
recordio_newreader(path)
}
\arguments{
\item{path}{path to recordio file.}
}
\value{
a reader object
}
\description{
Create a recordio reader.
}
\examples{
f <- recordio_newreader("/tmp/test.recordio")
data <- recordio_next(f)
recordio_close(f)
}
