% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recordio.R
\name{recordio_next}
\alias{recordio_next}
\title{Read the next record from a recordio reader}
\usage{
recordio_next(r)
}
\arguments{
\item{r}{reader object created by recordio_newreader.}
}
\value{
a raw byte vector, or nil in case of EOF.
}
\description{
Read the next record from a recordio reader
}
